+++ BackEnd/requirements.txt
@@ -10,18 +10,22 @@ chardet==4.0.0
 click==7.1.2
 colorama==0.4.4
 coverage==5.4
+cryptography==3.4.4
 dicttoxml==1.7.4
 dominate==2.6.0
 feedformatter @ https://github.com/marianoguerra/feedformatter/archive/master.zip
 Flask==1.1.2
 Flask-Bootstrap4==4.0.2
 Flask-Breadcrumbs==0.5.1
+flask-core==2.9.0
+Flask-Cors==3.0.10
 Flask-JWT-Extended==3.25.0
 Flask-Login==0.5.0
 Flask-Menu==0.7.2
 Flask-PyMongo==2.3.0
 flask-restx==0.2.0
 Flask-SocketIO==5.0.1
+Flask-SQLAlchemy==2.4.4
 Flask-WTF==0.14.3
 gevent==21.1.2
 gevent-websocket==0.10.1
@@ -47,6 +51,8 @@ nltk==3.5
 numpy==1.20.0
 packaging==20.9
 pluggy==0.13.1
+psycopg2==2.8.6
+psycopg2-binary==2.8.6
 py==1.10.0
 pycparser==2.20
 PyJWT==1.7.1
@@ -66,6 +72,7 @@ requirements-parser==0.2.0
 six==1.15.0
 sleekxmpp==1.3.3
 soupsieve==2.1
+SQLAlchemy==1.3.23
 tempora==4.0.1
 toml==0.10.2
 tqdm==4.56.0
